requirements.txt
huggingface-hub
llama-cpp-python
streamlit